+++ src/lib/xurl.test.ts
@@ -18,7 +18,8 @@ const FOLLOW_USER_FIELDS =
 	"id%2Cusername%2Cname%2Cdescription%2Cverified%2Cprotected%2Cpublic_metrics%2Cprofile_image_url%2Ccreated_at";
 const LIST_FIELDS =
 	"created_at%2Cdescription%2Cfollower_count%2Cmember_count%2Cname%2Cowner_id%2Cprivate";
-const AUTHOR_MEDIA_EXPANSIONS = "author_id%2Cattachments.media_keys";
+const AUTHOR_MEDIA_EXPANSIONS =
+	"author_id%2Cattachments.media_keys%2Creferenced_tweets.id%2Creferenced_tweets.id.author_id";
 const MEDIA_EXPANSION = "attachments.media_keys";
 const MEDIA_FIELDS =
 	"variants%2Cpreview_image_url%2Curl%2Cduration_ms%2Calt_text%2Ctype%2Cwidth%2Cheight%2Cpublic_metrics";